backend/requirements.txt
fastapi==0.111.1
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
mysql-connector-python===9.0.0